Merge in CGAT/gpssample from mapbox3 to Development

* commit '40c68d1cd815dbb9b4aa9fc9607fc6693fc60594':
  Current location updates
  Added current location to CreateEnumerationAreaFragment and PerformEnumerationFragment.
  migrated PerformEnumerationFragment to MapBox
  updated string translations
  migrated CreateConfigurationFragment to Mapbox
  migrated ConfigurationFragment to Mapbox
  migrated CreateEnumerationAreaFragment to Mapbox
  Added mapbox to project

+++ app/src/main/java/edu/gtri/gpssample/viewmodels/models/PerformEnumerationModel.kt
@@ -5,11 +5,11 @@ import androidx.lifecycle.MutableLiveData
 
 class PerformEnumerationModel
 {
-    private var _currentZoomLevel : MutableLiveData<Float>? = null
+    private var _currentZoomLevel : MutableLiveData<Double>? = null
 
-    var currentZoomLevel : LiveData<Float>? = _currentZoomLevel
+    var currentZoomLevel : LiveData<Double>? = _currentZoomLevel
 
-    fun setCurrentZoomLevel( zoomLevel: Float )
+    fun setCurrentZoomLevel( zoomLevel: Double )
     {
         _currentZoomLevel = MutableLiveData(zoomLevel)
         currentZoomLevel = _currentZoomLevel
